Create chagnes

+++ server/controller/blogController.js
@@ -60,15 +60,24 @@ export const updateBlog = async (req, res) => {
   }
 };
 
+import mongoose from "mongoose";
+
 export const deleteBlog = async (req, res) => {
   try {
-    const id = req.params.id;
+    const { id } = req.params;
+
+    // Proverite da li je ID validan
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid blog ID" });
+    }
+
     const singlePost = await modelBlog.findById(id);
     if (!singlePost) {
-      return res.status(404).json({ message: "Single post doesn't existing" });
+      return res.status(404).json({ message: "Blog post doesn't exist" });
     }
+
     await modelBlog.findByIdAndDelete(id);
-    res.status(200).json({ message: "User deleted succesufully" });
+    res.status(200).json({ message: "Blog post deleted successfully" });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
